As far as I can tell, this is defined to be an off_t, not a size_t.
git-svn-id: https://svn.parrot.org/parrot/trunk@2498 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ src/embed.c
@@ -68,7 +68,7 @@ Parrot_setwarnings(struct Parrot_Interp *interpreter, Parrot_warnclass wc)
 struct PackFile *
 Parrot_readbc(struct Parrot_Interp *interpreter, char *filename)
 {
-    size_t program_size;
+    off_t program_size;
     char *program_code;
     struct PackFile *pf;
     ParrotIO * io = NULL;
